Default Dark Mode: Expansion (pack.mcmeta)
I haven’t updated the pack.mcmeta on GitHub, so I’ve now done it here. For anyone downloading the GitHub version, it will always be labeled as “2025.X.X” to show it’s the GitHub release. The year will change as needed, but this helps me identify the GitHub version better. For example, Modrinth/CurseForge releases will continue to be dated “2025.1.0”. based on the actual date as usual.